requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
python-decouple==3.8
